--- study-juc/src/main/java/org/ohx/studyjuc/completablefuture/README.md
@@ -1,1 +1,0 @@
-CompletableFuture 的学习 Demo
