--- study-juc/src/main/java/org/ohx/studyjuc/completablefuture/basedemo/Demo.java
@@ -1,17 +1,0 @@
-package org.ohx.studyjuc.completablefuture.basedemo;
-
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ExecutionException;
-
-/**
- * @author mudkip
- * @date 2022/7/14
- */
-public class Demo {
-    public static void main(String[] args) throws ExecutionException, InterruptedException {
-        CompletableFuture<Void> completableFuture = CompletableFuture.runAsync(() -> {
-            System.out.println(Thread.currentThread().getName());
-        });
-        System.out.println(completableFuture.get());
-    }
-}
