+++ study-juc/src/main/java/org/ohx/studyjuc/countdownlatch/README.md
@@ -1,3 +1,5 @@
+# CountDownLatch 的学习 Demo
+
 Latch中文翻译是门闩，也就是有“门锁”的功能，所以当门没有打开时，N个人是不能进入屋内的，也就是N个线程是不能继续向下运行的，支持这样的特性可以控制线程执行任务的时机，使线程以“组团”的方式一起执行任务。
 
 CountDownLatch类是一个同步功能辅助类，功能是给定一个计数，当使用这个类的线程判断计数不为0时，则呈wait状态；如果计数为0则继续执行。支持这样的特性可以控制线程执行任务的时机，使线程以“组团”的方式一起执行任务。
