--- study-juc/src/main/java/org/ohx/studyjuc/countdownlatch/test2/Run.java
@@ -1,30 +1,0 @@
-package org.ohx.studyjuc.countdownlatch.test2;
-
-import java.util.Random;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
-
-/**
- * @author mudkip
- * @date 2022/7/12
- */
-public class Run {
-    public static void main(String[] args) throws InterruptedException {
-        CountDownLatch countDownLatch = new CountDownLatch(10);
-
-        for (int i = 0; i < 10; i++) {
-            new Thread(() -> {
-                try {
-                    TimeUnit.SECONDS.sleep(new Random().nextInt(5));
-                    System.out.println(Thread.currentThread().getName() + " finish!");
-                    countDownLatch.countDown();
-                } catch (InterruptedException e) {
-                    e.printStackTrace();
-                }
-            }, i + 1 + "").start();
-        }
-
-        countDownLatch.await();
-        System.out.println("计数为0，继续执行。");
-    }
-}
