+++ study-juc/src/main/java/org/ohx/studyjuc/cyclicbarrier/demo1/Run.java
@@ -1,0 +1,28 @@
+package org.ohx.studyjuc.cyclicbarrier.demo1;
+
+import java.util.Random;
+import java.util.concurrent.BrokenBarrierException;
+import java.util.concurrent.CyclicBarrier;
+import java.util.concurrent.TimeUnit;
+
+/**
+ * @author mudkip
+ * @date 2022/7/12
+ */
+public class Run {
+    public static void main(String[] args) {
+        CyclicBarrier cyclicBarrier = new CyclicBarrier(2, () -> System.out.println("全都到了！"));
+
+        for (int i = 0; i < 10; i++) {
+            new Thread(() -> {
+                try {
+                    TimeUnit.SECONDS.sleep(new Random().nextInt(10));
+                    System.out.println(Thread.currentThread().getName() + "到了！");
+                    cyclicBarrier.await();
+                } catch (InterruptedException | BrokenBarrierException e) {
+                    e.printStackTrace();
+                }
+            }, i + 1 + "").start();
+        }
+    }
+}
