+++ study-juc/src/main/java/org/ohx/studyjuc/join/README.md
@@ -1,6 +1,16 @@
+# Thread.join() 的学习 Demo
+
 在很多情况下，主线程创建并启动子线程，如果子线程要进行大量的耗时运算，主线程往往将早于子线程结束之前结束，这时如果主线程想等待子线程执行完成之后再结束，例如，当子线程处理一个数据，主线程要取得这个数据中的值时，就要用到 join()
 方法了。方法 join() 的作用是等待线程对象销毁。
 
 join() 方法的作用是使所属的线程对象x正常执行 run() 方法中的任务，而使当前线程z进行无限期的阻塞，等待线程x销毁后再继续执行线程z后面的代码，具有串联执行的效果。
 
 join()方法具有使线程排队运行的效果，有些类似同步的运行效果，但是join()方法与synchronized的区别是join()方法在内部使用wait()方法进行等待，而synchronized关键字使用锁作为同步。
+
+## demo1
+
+基本使用
+
+## demo2
+
+使用 join() 方法要注意坑
