+++ study-juc/src/main/java/org/ohx/studyjuc/join/demo2/Run1.java
@@ -1,0 +1,18 @@
+package org.ohx.studyjuc.join.demo2;
+
+/**
+ * @author mudkip
+ * @date 2022/6/25
+ */
+public class Run1 {
+    public static void main(String[] args) throws InterruptedException {
+        ThreadB threadB = new ThreadB();
+        ThreadA threadA = new ThreadA(threadB);
+        threadA.start();
+        threadB.start();
+        // join方法需要抢ThreadB对象锁
+        threadB.join(200);
+
+        System.out.println("main end " + System.currentTimeMillis());
+    }
+}
