--- study-juc/src/main/java/org/ohx/studyjuc/join/jointest2/README.md
@@ -1,1 +1,0 @@
-使用 join() 方法要注意坑
